Refactored synchronization services to enhance separation of responsability and to facilitate code understanding
